+++ test/plugin/validation_helpers_test.rb
@@ -9,16 +9,28 @@ describe Shrine::Plugins::ValidationHelpers do
   describe "#validate_max_size" do
     it "adds an error if file is larger than given size" do
       @attacher.assign(fakeio("image"))
-      @attacher.validate_max_size 1
+      assert_equal false, @attacher.validate_max_size(1)
       refute_empty @attacher.errors
+    end
+
+    it "doesn't add an error if file is in the size limits" do
+      @attacher.assign(fakeio("image"))
+      assert_equal true, @attacher.validate_max_size(50)
+      assert_empty @attacher.errors
     end
   end
 
   describe "#validate_min_size" do
     it "adds an error if file is smaller than given size" do
       @attacher.assign(fakeio("image"))
-      @attacher.validate_min_size 10
+      assert_equal false, @attacher.validate_min_size(10)
       refute_empty @attacher.errors
+    end
+
+    it "doesn't add an error if file is in the size limits" do
+      @attacher.assign(fakeio("image"))
+      assert_equal true, @attacher.validate_min_size(1)
+      assert_empty @attacher.errors
     end
   end
 
@@ -26,20 +38,27 @@ describe Shrine::Plugins::ValidationHelpers do
     it "adds an error if file is wider than given size" do
       @attacher.shrine_class.plugin :store_dimensions
       @attacher.assign(image)
-      @attacher.validate_max_width 10
+      assert_equal false, @attacher.validate_max_width(10)
       refute_empty @attacher.errors
     end
 
-    it "doesn't trigger if width is missing" do
+    it "doesn't add an error if file is in the dimension limits" do
+      @attacher.shrine_class.plugin :store_dimensions
+      @attacher.assign(image)
+      assert_equal true, @attacher.validate_max_width(500)
+      assert_empty @attacher.errors
+    end
+
+    it "doesn't add an error if width is nil" do
       @attacher.shrine_class.plugin :store_dimensions
       @attacher.assign(fakeio("not an image"))
-      @attacher.validate_max_width 10
+      assert_equal nil, @attacher.validate_max_width(10)
       assert_empty @attacher.errors
     end
 
     it "requires the store_dimensions plugin" do
       @attacher.assign(image)
-      assert_raises(Shrine::Error) { @attacher.validate_max_width 500 }
+      assert_raises(Shrine::Error) { @attacher.validate_max_width(500) }
     end
   end
 
@@ -47,20 +66,27 @@ describe Shrine::Plugins::ValidationHelpers do
     it "adds an error if file is narrower than given size" do
       @attacher.shrine_class.plugin :store_dimensions
       @attacher.assign(image)
-      @attacher.validate_min_width 500
+      assert_equal false, @attacher.validate_min_width(500)
       refute_empty @attacher.errors
     end
 
-    it "doesn't trigger if width is missing" do
+    it "doesn't add an error if file is in the dimension limits" do
+      @attacher.shrine_class.plugin :store_dimensions
+      @attacher.assign(image)
+      assert_equal true, @attacher.validate_min_width(10)
+      assert_empty @attacher.errors
+    end
+
+    it "doesn't add an error if width is nil" do
       @attacher.shrine_class.plugin :store_dimensions
       @attacher.assign(fakeio("not an image"))
-      @attacher.validate_min_width 10
+      assert_equal nil, @attacher.validate_min_width(10)
       assert_empty @attacher.errors
     end
 
     it "requires the store_dimensions plugin" do
       @attacher.assign(image)
-      assert_raises(Shrine::Error) { @attacher.validate_min_width 10 }
+      assert_raises(Shrine::Error) { @attacher.validate_min_width(10) }
     end
   end
 
@@ -68,20 +94,27 @@ describe Shrine::Plugins::ValidationHelpers do
     it "adds an error if file is wider than given size" do
       @attacher.shrine_class.plugin :store_dimensions
       @attacher.assign(image)
-      @attacher.validate_max_height 10
+      assert_equal false, @attacher.validate_max_height(10)
       refute_empty @attacher.errors
     end
 
-    it "doesn't trigger if height is missing" do
+    it "doesn't add an error if file is in the dimension limits" do
+      @attacher.shrine_class.plugin :store_dimensions
+      @attacher.assign(image)
+      assert_equal true, @attacher.validate_max_height(500)
+      assert_empty @attacher.errors
+    end
+
+    it "doesn't add an error if height is missing" do
       @attacher.shrine_class.plugin :store_dimensions
       @attacher.assign(fakeio("not an image"))
-      @attacher.validate_max_height 10
+      assert_equal nil, @attacher.validate_max_height(10)
       assert_empty @attacher.errors
     end
 
     it "requires the store_dimensions plugin" do
       @attacher.assign(image)
-      assert_raises(Shrine::Error) { @attacher.validate_max_height 500 }
+      assert_raises(Shrine::Error) { @attacher.validate_max_height(500) }
     end
   end
 
@@ -89,31 +122,38 @@ describe Shrine::Plugins::ValidationHelpers do
     it "adds an error if file is narrower than given size" do
       @attacher.shrine_class.plugin :store_dimensions
       @attacher.assign(image)
-      @attacher.validate_min_height 500
+      assert_equal false, @attacher.validate_min_height(500)
       refute_empty @attacher.errors
     end
 
-    it "doesn't trigger if height is missing" do
+    it "doesn't add an error if file is in the dimension limits" do
+      @attacher.shrine_class.plugin :store_dimensions
+      @attacher.assign(image)
+      assert_equal true, @attacher.validate_min_height(1)
+      assert_empty @attacher.errors
+    end
+
+    it "doesn't add an error if height is nil" do
       @attacher.shrine_class.plugin :store_dimensions
       @attacher.assign(fakeio("not an image"))
-      @attacher.validate_min_height 10
+      assert_equal nil, @attacher.validate_min_height(10)
       assert_empty @attacher.errors
     end
 
     it "requires the store_dimensions plugin" do
       @attacher.assign(image)
-      assert_raises(Shrine::Error) { @attacher.validate_min_height 10 }
+      assert_raises(Shrine::Error) { @attacher.validate_min_height(10) }
     end
   end
 
   describe "#validate_mime_type_inclusion" do
     it "adds an error when mime_type is not in the whitelist" do
       @attacher.assign(fakeio(content_type: "video/mpeg4"))
-      @attacher.validate_mime_type_inclusion ["image/jpeg", "image/png"]
+      assert_equal false, @attacher.validate_mime_type_inclusion(["image/jpeg", "image/png"])
       refute_empty @attacher.errors
 
       @attacher.assign(fakeio(content_type: "image/jpeg"))
-      @attacher.validate_mime_type_inclusion ["image/jpeg", "image/png"]
+      assert_equal true, @attacher.validate_mime_type_inclusion(["image/jpeg", "image/png"])
       assert_empty @attacher.errors
     end
 
@@ -135,7 +175,7 @@ describe Shrine::Plugins::ValidationHelpers do
 
     it "adds an error if mime_type is missing" do
       @attacher.assign(fakeio)
-      @attacher.validate_mime_type_inclusion [/image/]
+      assert_equal false, @attacher.validate_mime_type_inclusion([/image/])
       refute_empty @attacher.errors
     end
   end
@@ -143,11 +183,11 @@ describe Shrine::Plugins::ValidationHelpers do
   describe "#validate_mime_type_exclusion" do
     it "adds an error when mime_type is in the blacklist" do
       @attacher.assign(fakeio(content_type: "video/mpeg4"))
-      @attacher.validate_mime_type_exclusion ["video/mpeg4", "audio/mp3"]
+      assert_equal false, @attacher.validate_mime_type_exclusion(["video/mpeg4", "audio/mp3"])
       refute_empty @attacher.errors
 
       @attacher.assign(fakeio(content_type: "image/jpeg"))
-      @attacher.validate_mime_type_exclusion ["video/mpeg4", "audio/mp3"]
+      assert_equal true, @attacher.validate_mime_type_exclusion(["video/mpeg4", "audio/mp3"])
       assert_empty @attacher.errors
     end
 
@@ -163,7 +203,7 @@ describe Shrine::Plugins::ValidationHelpers do
 
     it "doesn't add an error if mime_type is missing" do
       @attacher.assign(fakeio)
-      @attacher.validate_mime_type_exclusion [/video/]
+      assert_equal true, @attacher.validate_mime_type_exclusion([/video/])
       assert_empty @attacher.errors
     end
   end
@@ -171,11 +211,11 @@ describe Shrine::Plugins::ValidationHelpers do
   describe "#validate_extension_inclusion" do
     it "adds an error when extension is not in the whitelist" do
       @attacher.assign(fakeio(filename: "video.mp4"))
-      @attacher.validate_extension_inclusion ["jpg", "png"]
+      assert_equal false, @attacher.validate_extension_inclusion(["jpg", "png"])
       refute_empty @attacher.errors
 
       @attacher.assign(fakeio(filename: "image.jpg"))
-      @attacher.validate_extension_inclusion ["jpg", "png"]
+      assert_equal true, @attacher.validate_extension_inclusion(["jpg", "png"])
       assert_empty @attacher.errors
     end
 
@@ -197,7 +237,7 @@ describe Shrine::Plugins::ValidationHelpers do
 
     it "adds an error if extension is missing" do
       @attacher.assign(fakeio)
-      @attacher.validate_extension_inclusion [/jpg/]
+      assert_equal false, @attacher.validate_extension_inclusion([/jpg/])
       refute_empty @attacher.errors
     end
   end
@@ -205,11 +245,11 @@ describe Shrine::Plugins::ValidationHelpers do
   describe "#validate_extension_exclusion" do
     it "adds an error when extension is in the blacklist" do
       @attacher.assign(fakeio(filename: "video.mp4"))
-      @attacher.validate_extension_exclusion ["mp4", "mp3"]
+      assert_equal false, @attacher.validate_extension_exclusion(["mp4", "mp3"])
       refute_empty @attacher.errors
 
       @attacher.assign(fakeio(filename: "image.jpg"))
-      @attacher.validate_extension_exclusion ["mp4", "mp3"]
+      assert_equal true, @attacher.validate_extension_exclusion(["mp4", "mp3"])
       assert_empty @attacher.errors
     end
 
@@ -231,7 +271,7 @@ describe Shrine::Plugins::ValidationHelpers do
 
     it "doesn't add an error if extension is missing" do
       @attacher.assign(fakeio)
-      @attacher.validate_extension_exclusion [/mp4/]
+      assert_equal true, @attacher.validate_extension_exclusion([/mp4/])
       assert_empty @attacher.errors
     end
   end