+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Make validation methods in `validation_helpers` plugin return whether validation succeeded (@janko-m)
+
 * Make extension matching case insensitive in `validation_helpers` plugin (@jonasheinrich)
 
 * Make `remove_invalid` plugin remove dirty state on attacher after removing invalid file (@janko-m)